chore: delete flow case and add up to one
