+++ cypress/e2e/authentication/logout.cy.ts
